ClassSignatureComparerHelper class
